requirements.txt
python==3.12.4
python-dotenv
PyYAML
openai